Don't re-finish GB init if recreating Activity
GutenbergEditorFragment has been recreated but the GB container
hasn't so, no need to re-send content, etc. Essentially, we do let the
system kill and recreate the fragments, and only let the parent continue
with the "setContent()" logic if the GB container was in fact
re-initialized.

That handles the cases of the system killing the activities (like when
"Don't keep activities" developer setting is enabled, and when rotating
the device.